requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.9
decorator==4.1.2
defusedxml==0.5.0
Django==1.11.6
django-allauth==0.33.0
django-filter==1.1.0
django-mptt==0.8.7
djangorestframework==3.7.1
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
Markdown==2.6.9
oauthlib==2.0.4
parso==0.1.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
Pygments==2.2.0
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
win-unicode-console==0.5
